Move grouping logic to TableUtils and use same pattern on PlainTableOutput
