+++ packages/ckeditor5-table/src/converters/downcast.ts
@@ -35,30 +35,9 @@ export function downcastTable( tableUtils: TableUtils, options: DowncastTableOpt
 			);
 		}
 
-		// Group rows.
-		const tableRowElements = Array.from( table.getChildren() )
-			.filter( child => child.is( 'element', 'tableRow' ) && child.index! >= headingRows );
-		const rowGroupMap = new Map();
-		let currentGroup = 0;
-
-		tableRowElements.forEach( trElement => {
-			// Ensure a rowGroup exists.
-			// This is a workaround / idea - currently if a row is inserted, it won't have the rowGroup attribute.
-			// Currently just putting it into the previous row group, although this won't always be intentional.
-			// If this approach to row groups in general is accepted this could be resolved in tableUtils.insertRows by
-			// copying the rewGroup from the row used as reference.
-			currentGroup = trElement.hasAttribute( 'rowGroup' ) ? Number( trElement.getAttribute( 'rowGroup' ) ) : currentGroup;
-			trElement._setAttribute( 'rowGroup', currentGroup );
-
-			if ( !rowGroupMap.has( currentGroup ) ) {
-				rowGroupMap.set( currentGroup, [] );
-			}
-
-			rowGroupMap.get( currentGroup ).push( trElement );
-		} );
-
 		// Table body slots.
 		if ( headingRows < tableUtils.getRows( table ) ) {
+			const rowGroupMap = tableUtils.getGroupedRows( table );
 			for ( const key of rowGroupMap.keys() ) {
 				writer.insert(
 					writer.createPositionAt( tableElement, 'end' ),
